tests: migrate issue-56 diagnostics into maintained pytest suite
- New tests/test_joint_limits_prior.py owns the LimitPrior read path, the
  wide-open fallback, joint_limits validation, the fitter hinge-loss
  guarantees (zero in-range / positive + corrective gradient on violation)
  and the neural penalty checks (marked slow: heavy regressor imports).
- Drop diagnostics/ one-off scripts: results are recorded in
  docs/design/issue56_implementation.md; check_remap_math.py duplicated the
  membership property tests/test_axis_remap.py already covers;
  posedirs_validation.py / blender_posedirs_check.py validated the #24
  change, which is merged upstream (#89).
